% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnoVaOMICS.R
\name{AnoVaOMICS}
\alias{AnoVaOMICS}
\title{Automatic calculation of an ANOVA for OMICS data}
\usage{
AnoVaOMICS(df, grp, threshold = 5, method.Adj = "BH")
}
\arguments{
\item{df}{dataframe (variable in columns, samples in rows)}

\item{grp}{sample grouping column}

\item{threshold}{threshold for significance (expressed as a percentage)}

\item{method.Adj}{method used for multiple comparison adjustment ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}
}
\value{
dataframe with significant variables
}
\description{
Automatic calculation of an ANOVA for OMICS data
}
\examples{
data(LIHC_pnas)
# subset of TCGA data from LIHC
res<-AnoVaOMICS(LIHC_pnas, iHCC, threshold=1) # 1\% threshold
}
