% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncrAUC.R
\name{IncrAUC}
\alias{IncrAUC}
\title{Incremental AUC}
\usage{
IncrAUC(df, grp, graph = T)
}
\arguments{
\item{df}{dataframe (measure times in minutes in columns, samples in rows)}

\item{grp}{sample grouping column}

\item{graph}{if TRUE, plot a boxplot, Default: TRUE}
}
\value{
the function returns a dataframe with the AUC per sample and a boxplot (if graph=T)
}
\description{
This function allows to calculate the incremental AUC (IAUC) per sample
for an ITT (Insulin Tolerance Test) or a GTT (Glucose Tolerance Test).
}
\examples{
data(OGTT)
# Artificial data from an OGTT experiment in KO and WT animals
IncrAUC(OGTT, lineage)
}
\seealso{
Brouns et al. 2005 (Nutrition Research Reviews)

Wolever et al. 2004 (Diabetes Care)
}
