% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.rm.R
\name{NA.rm}
\alias{NA.rm}
\title{Delete samples or variables with predefined number of NA}
\usage{
NA.rm(df, n = 0, direction = "rows")
}
\arguments{
\item{df}{dataframe}

\item{n}{maximum number of NA (if 0 --> all NAs)}

\item{direction}{in which direction the deletion will take place ("rows" or "columns")}
}
\value{
dataframe after sorting and deleting NA
}
\description{
Delete samples or variables with predefined number of NA
}
\examples{
#' # Simulated data via rnorm function
df<-as.data.frame(matrix(rnorm(40*20, mean=0, sd=1),40,20))
set.seed(1234)
df<-as.data.frame(lapply(df, function(cc) cc[sample(c(TRUE, NA),
                                                  prob = c(0.85, 0.15),
                                                   size = length(cc),
                                                     replace = TRUE) ]))
df2 <- NA.rm(df, 8, direction="columns")
head(df[,1:5]) # before
head(df2[,1:5]) # after
# Variable V2 removed
}
