% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxBatch.R
\name{wilcoxBatch}
\alias{wilcoxBatch}
\title{Automatic calculation of a Wilcoxon test for OMICS data}
\usage{
wilcoxBatch(df, groupVar, threshold = 5, method.Adj = "BH")
}
\arguments{
\item{df}{dataframe (variable in columns, samples in rows)}

\item{groupVar}{sample grouping column}

\item{threshold}{threshold for significance (expressed as a percentage)}

\item{method.Adj}{method used for multiple comparison adjustment ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}
}
\value{
dataframe with significant variables
}
\description{
Automatic calculation of a Wilcoxon test for OMICS data
}
\examples{
data(LIHC_pnas)
# subset of TCGA data from LIHC
LIHC_pnas <- subset(LIHC_pnas, iHCC != "iHCC3")
res<-wilcoxBatch(LIHC_pnas, iHCC, threshold=1) # 1\% threshold
}
