% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.rm.R
\name{NA.rm}
\alias{NA.rm}
\title{Delete samples or variables with predefined number of NA}
\usage{
NA.rm(df, n = 0, direction = "rows")
}
\arguments{
\item{df}{dataframe}

\item{n}{number of NA (if 0 --> all NAs)}

\item{direction}{in which direction the deletion will take place ("rows" or "columns")}
}
\value{
dataframe after sorting and deleting NA
}
\description{
Delete samples or variables with predefined number of NA
}
\examples{
data(simulate_data1)
# Simulated data via rnorm function
df2 <- NA.rm(simulate_data1, 2, direction="rows")
head(simulate_data1[,1:5]) # before
head(df2[,1:5]) # after
# Sample 5 removed
}
